+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+[1.7.0] 21.06.2026
+
+- [FAM](https://ticktick.com/webapp/#p/683eff21ebb75c000000031c/tasks/6a29d76febb75c0000000522): Delete transaction confirm
+- [FAM](https://ticktick.com/webapp/#p/683eff21ebb75c000000031c/tasks/6a341da3ebb75c000000035a): Transfer transaction, add new amount
+- [FAM](https://ticktick.com/webapp/#p/683eff21ebb75c000000031c/tasks/6a29d68cebb75c0000000483): Mobile AI 2.0
+
 [1.6.0] 11.06.2026
 
 - [FAM](https://ticktick.com/webapp/#p/683eff21ebb75c000000031c/tasks/6a29d6a7ebb75c000000048b): Normalise balance